# packages/backend/src/swagger.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/apps/{urn}": {
      "get": {
        "operationId": "getApp",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAppDto"
                }
              }
            }
          }
        },
        "tags": [
          "Apps"
        ]
      }
    },
    "/api/apps/{urn}/compose-diff": {
      "get": {
        "operationId": "getAppComposeDiff",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetComposeDiffDto"
                }
              }
            }
          }
        },
        "tags": [
          "Apps"
        ]
      }
    },
    "/api/apps/{urn}/config-diff": {
      "get": {
        "operationId": "getAppConfigDiff",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetConfigDiffDto"
                }
              }
            }
          }
        },
        "tags": [
          "Apps"
        ]
      }
    },
...
    "/api/apps/{urn}/check-availability": {
      "get": {
        "operationId": "checkAvailability",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Apps"
        ]
      }
    },
...
    "/api/apps/{urn}/resolve-availability": {
      "post": {
        "operationId": "resolveAvailability",
        "parameters": [
          {
            "name": "urn",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Apps"
        ]
      }
    },
...
  },
  "info": {
    "title": "CI Hub API",
    "description": "API specs for CI Hub",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "GetAppDto": {
        "type": "object",
        "properties": {}
      },
      "GetComposeDiffDto": {
        "type": "object",
        "properties": {}
      },
      "GetConfigDiffDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}